# src/main/java/com/eaglebank/controller/AuthController.java
package com.eaglebank.controller;

import com.eaglebank.dto.Response;
import com.eaglebank.dto.auth.AuthRequest;
import com.eaglebank.dto.auth.AuthResponse;
import com.eaglebank.dto.error.BadRequestErrorResponse;
import com.eaglebank.dto.error.ErrorDetail;
import com.eaglebank.service.AuthService;
import jakarta.validation.Valid;
import org.slf4j.Logger;
import org.slf4j.LoggerFactory;
import org.springframework.http.HttpStatus;
import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

import java.util.ArrayList;

@RestController
@RequestMapping("/v1/auth")
public class AuthController {

    private static final Logger log = LoggerFactory.getLogger(AuthController.class);
    private final AuthService authService;

    public AuthController(AuthService authService) {
        this.authService = authService;
    }

    @PostMapping("/login")
    public ResponseEntity<Response> authenticateUser(@Valid @RequestBody AuthRequest request) {
        try {
            AuthResponse authResponse = authService.authenticateUser(request);
            return ResponseEntity.ok(authResponse);
        }catch (Exception e){
            var errorList = new ArrayList<ErrorDetail>();
            errorList.add(new ErrorDetail(null, e.getMessage(), null));
            return new ResponseEntity<>(new BadRequestErrorResponse("Unauthorized",
                    errorList), HttpStatus.UNAUTHORIZED);
        }
    }
}