# src/main/java/com/eaglebank/dto/account/BankAccountResponse.java
package com.eaglebank.dto.account;

import com.eaglebank.entity.Account;
import lombok.Data;

import java.time.LocalDateTime;

@Data
public class BankAccountResponse {
    private String accountNumber;
    private String sortCode;
    private String name;
    private Account.AccountType accountType;
    private double balance;
    private Account.Currency currency;
    private String userId;
    private LocalDateTime createdTimestamp;
    private LocalDateTime updatedTimestamp;
}