# src/main/java/com/eaglebank/dto/error/ErrorResponse.java
package com.eaglebank.dto.error;

import com.eaglebank.dto.Response;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.EqualsAndHashCode;
import lombok.NoArgsConstructor;

@EqualsAndHashCode(callSuper = true)
@Data
@NoArgsConstructor
@AllArgsConstructor
public class ErrorResponse extends Response {
    private String message;
}